Automatically setup the hirte-agent
Attempt to first copy the /etc/hirte/agent.conf from
the host into the hirte environmnet, and add qm. to the
front of the NodeName field.

If the NodeName field does not exists or is emtpy, then add
qm.HOSTNAME as the NodeName.

If the NodeName is qm., then add HOSTNAME to the end.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>