Update qm /etc/agent/agent.conf that when the qm.service is started
This pull request will cause the /usr/lib/qm/rootfs/etc/hirte/agent.conf
to be created on every qm.service start.  This way if the admin changes
the /etc/hirte/agent.conf on the host, the equivalent inside of the qm
will match with the exception that the nodename is prefixed with qm.

If the nodename in the hosts /etc/hirte/agent.conf is changed to foobar,
then the qm's /etc/hirte/agent.conf will have the same settings except
the node name will be qm.foobar.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>