Fix handling of /var and /etc within container
With the current code using builtin volumes users on the
host system can not modify the ROOTFS/var and ROOTFS/etc
directories.
This PR just remounts these directories on top of themselves
read/write allowing users from the host to modify the
directories and the processes inside of the container to modify
them.
Both the host and the QM will be able to see the modifications.

Added a ROOTFS=/usr/lib/qm/rootfs environment variable to make
the quadlet easier to understand.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>