qm.container: replace tz=local in ostree distro
In recent tests with ostree + quadlet environment the QM service was not able
to work correctly, podman failed to remove /etc/localtime affecting the QM
service. This patch is a workaround to Timezone=local which uses
Environment=TZ to set the default timezone to UTC until we work in root cause.

More information: #367

Fixes: #367

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>