seccomp: use SYSCALLS_TO_DENY array
This global array can be loaded externally into tools to check
if programs can be added into QM image/yaml that are not allowed
or won't work into the QM partition.

In this case: "sched_setscheduler" and "sched_setattr" which QM
engineers are following the Risk Assessments rules.

Projects like CentOS Automotive Stream Distribution could be take
advanced in the CI/CD pipeline.

Finally, added an example in tool, how could be used in a CI/CD
scenario such integration.

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>