qm.if: add rules for wayland support
Add minimal set of SELinux rules to allow
running a wayland compositor run as a
container workload inside QM.

In this first iteration, rules assume
using the QM system_bus_socket for the
dbus server. Next step is to allow
a QM-specific dbus daemon to run with
its own session and socket.

Signed-off-by: Albert Esteve <aesteve@redhat.com>