tools/version-update: script to manage version
Create a script to ease version managing for
maintainers and developers. The script will handle
updating the version in all required files
consistently.

By default, if no new version is provided, the
script increases the minor version by 1.
Otherwise, a specific version can be set with:

$ tools/version-update --version 1.0

And the tools will just perform the changes in
all the relevant files.

Signed-off-by: Albert Esteve <aesteve@redhat.com>