Fix suport for CPUWeight
In the current setup, qm.service is part of the QM.slice. This means
that the CPUWeight of qm.service is only applied against the other
groups of QM.slice, i.e. none. This means CPUWeight does nothing.

Here change the parent of qm.service to be -.service (the root
cgroup), which means that the weight of it will be compared to the
other toplevel slices, which normally is user.slice and system.slice.

So, with the current default CPUWeight in qm.container of 50, and the
default weight of 100, this means that qm.service now will get half
the cpu of the processes in system.slice, which is the intent.

I tried this by running "openssl speed --seconds 99999" both inside qm
and outside it one a one-cpu VM, and before both processes always got
50% cpu, but now they are split up according to CPUWeight in the
qm.container.

Signed-off-by: Alexander Larsson <alexl@redhat.com>