Remove application specific settings
The QM SELinux policy is not the right place to assign
application-specific rights. If there is a systematic problem with the
QM policy, it should be solved independently of the specific application
with which the problem was first identified.

However, if the problem should actually be solved upstream in the
application, it makes sense to document in the qm README what
constitutes good practice for adapting to the QM partition concept.

Signed-off-by: Sebastian Hetze <she@redhat.com>