wayland: add qm-wayland package
Add qm-wayland package. This is a revamp of the
previous windowmanager subpackage. But in this case,
it is not a qm subpackage in the sense of requiring QM
and installing in rootfs. Instead, this qm-wayland
package is meant to be installed within QM. It adds
the files required to run the wayland scenario, and
depends on container images maintained on quay.io.

This is the first draft, but it requires the
dev-binder-hook to function correctly.

Signed-off-by: Albert Esteve <aesteve@redhat.com>