Add qm-wayland package (#934)
* windowmanager: remove subpackage

Remove windowmanager subsystem package. We
want to revamp the scenario and allow the package
to be installed within QM instead of extending QM.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* wayland: add qm-wayland package

Add qm-wayland package. This is a revamp of the
previous windowmanager subpackage. But in this case,
it is not a qm subpackage in the sense of requiring QM
and installing in rootfs. Instead, this qm-wayland
package is meant to be installed within QM. It adds
the files required to run the wayland scenario, and
depends on container images maintained on quay.io.

This is the first draft, but it requires the
dev-binder-hook to function correctly.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* subsystems/wayland: add container support

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* packit: update files to build qm-wayland

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* qm-oci-hooks: make the package qm-independent

Make qm-oci-hooks package indepent from qm. It can then
be installed both in the host or within the qm to get
the hooks installed in the right folders.

As it is indepent we can now add it to the Packit
configuration so it gets built and deployed automatically.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* qm-oci-hooks: various fixes

Split seat and devices hooks as otherwise the
annotations were not matching correctly to execute
the hooks when they should.

Also, fix the package to add a missing dependency
to include udevadm application.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* qm-wayland: make it use qm-oci-hooks

Now that hooks is an independent package, qm-wayland
can recomment it to work properly.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* Makefile: exclude packages already in packit from subpackages

Exclude packages that were added to packit from the
subpackages target. That will exclude also from the
`Build Subpackages` job in the pipelines, as the build
environment have missing dependencies.

Signed-off-by: Albert Esteve <aesteve@redhat.com>

* tox.ini: make lint match pre-commit rules

Signed-off-by: Albert Esteve <aesteve@redhat.com>

---------

Signed-off-by: Albert Esteve <aesteve@redhat.com>